A refactoring in the internal Fetch API was introduced in Sidekiq 6.1.0 that is not backwards compatible.
As part of this refactoring the object in the :fetch option is expected
to be an instance of Sidekiq::BasicFetch or an object that responds to
the same methods.

This change allows using the gem with newer versions while keeping
compatibility with old ones. See https://github.com/mperham/sidekiq/pull/4602 for more details.